example/requirements.txt
arxiv==2.1.0
anthropic>=0.18.0
python-dotenv>=1.0.0
